+++ applications/PUfoam/Models/FoamConstruction/foamreconstr/src/geometry.hh
@@ -1,10 +1,10 @@
 #ifndef GEOMETRY_H
 #define GEOMETRY_H
 
-bool in_domain(float, float, float);
-float *CrossProduct(float *, float *);
-bool SameSide(float *,float *,float *,float *);
-bool PointInTriangle(float *,float *,float *,float *);
+bool in_domain(double, double, double);
+double *CrossProduct(double *, double *);
+bool SameSide(double *,double *,double *,double *);
+bool PointInTriangle(double *,double *,double *,double *);
 double porosity(int ***);
 
 #endif
