+++ applications/PUfoam/Models/FoamConstruction/foamreconstr/src/skeleton.hh
@@ -2,6 +2,6 @@
 #define SKELETON_H
 
 void makeFoamSkeleton(std::string, int, int *, int *, int *);
-void importFoamSkeleton(std::string, float **, int **, int, int, int &);
+void importFoamSkeleton(std::string, double **, int **, int, int, int &);
 
 #endif
